refactor : addRequired 위치 변경
