+++ src/component/AddLunchModalForm.js
@@ -18,6 +18,10 @@ const CATEGORY_ICON = {
   기타: "./category-etc.png",
 };
 
+export function addRequired(element, required) {
+  if (required) element.classList.add("form-item--required");
+}
+
 function AddLunchModalForm(restaurantList, modalId) {
   const $modalForm = toElement(`
     <form>
