+++ src/component/Header.js
@@ -1,4 +1,3 @@
-
 import toElement from "../utils/toElement.js";
 import append from "../utils/append.js";
 
@@ -7,7 +6,6 @@ function Header(iconButton) {
     <header class="gnb">
       <h1 class="gnb__title text-title">점심 뭐 먹지</h1>
     </header>`);
-
   append($header, iconButton);
 
   return $header;
