--- src/utils/required.js
@@ -1,5 +1,0 @@
-function addRequired(element, required) {
-  if (required) element.classList.add("form-item--required");
-}
-
-export default addRequired;
